+++ tests/run_examples.py
@@ -27,7 +27,7 @@ from time import gmtime, strftime
 
 from daal4py import __has_dist__
 from daal4py.sklearn._utils import get_daal_version
-from onedal.utils._dp_helpers import dpctl_available
+from onedal.utils._dpep_helpers import dpctl_available
 
 print("Starting examples validation")
 # First item is major version - 2021,
